# Optimization batch 7: use file basenames to guide rename detection