This series depends on ????.

This series uses file basenames in a basic fashion to guide rename
detection.  It represents "Optimization #3" from my Git Merge 2020
talk[1].  For the testcases mentioned in commit ??????????
("merge-ort: begin performance work; instrument with trace2_region_*
calls", 2020-10-28), the changes in just this series improves the
performance as follows:

                         Before Series           After Series
    no-renames:       13.815 s ±  0.062 s    13.138 s ±  0.086 s
    mega-renames:   1799.937 s ±  0.493 s   169.488 s ±  0.494 s
    just-one-mega:    51.289 s ±  0.019 s     5.061 s ±  0.017 s

As a reminder, before any merge-ort/diffcore-rename performance work,
the performance results we started with (as noted in the same commit
message) were:

    no-renames-am:      6.940 s ±  0.485 s
    no-renames:        18.912 s ±  0.174 s
    mega-renames:    5964.031 s ± 10.459 s
    just-one-mega:    149.583 s ±  0.751 s

[1] https://github.com/newren/presentations/blob/pdfs/merge-performance/merge-performance-slides.pdf